Expose environment variables of openvpn to iptables. Rearrange openvpn.
Fix: ping should not be of None value or else it can be sorted. Change it to
  float('inf')

Reset iptables to fallback when trying to connect.

Give openvpn connection a timeout of 10s and wait for environment scripts to
  finish before applying iptables rules files.

Try to connect 3 times in case of openvpn timeouts. Each time with the next best
  found server if not ping is 'inf'

Make changes work in daemon mode of openvpn.